Extract dashboard components and AI payment panel
