+++ src/pages/agent/AgentDashboard.jsx
@@ -3,359 +3,10 @@ import { useNavigate } from "react-router-dom";
 import MainLayout from "../../components/layout/Layout";
 import { AuthContext } from "../../context/AuthProvider";
 import api from "../../api/axios";
-
-// ─── Global CSS ───────────────────────────────────────────────────────────────
-const CSS = `
-  @import url('https://fonts.googleapis.com/css2?family=Cormorant+Garamond:ital,wght@0,400;0,600;0,700;1,400&family=DM+Sans:wght@300;400;500;600&display=swap');
-
-  .ad * { box-sizing: border-box; }
-  .ad { font-family: 'DM Sans', system-ui, sans-serif; background: #f2ede4; min-height: 100vh; }
-
-  .ad-card {
-    background: #faf7f2;
-    border: 1.5px solid #e8e2d6;
-    border-radius: 14px;
-    box-shadow: 0 2px 16px rgba(20,16,10,0.06);
-    overflow: hidden;
-  }
-
-  .ad-btn {
-    transition: all 0.17s ease;
-    cursor: pointer;
-    font-family: 'DM Sans', sans-serif;
-    border: none;
-  }
-  .ad-btn:hover { opacity: 0.85; transform: translateY(-1px); }
-
-  @keyframes ad-fade-up  { from{opacity:0;transform:translateY(18px)} to{opacity:1;transform:translateY(0)} }
-  @keyframes ad-scale-in { from{opacity:0;transform:scale(0.96)} to{opacity:1;transform:scale(1)} }
-  @keyframes ad-toast    { from{opacity:0;transform:translateY(10px)} to{opacity:1;transform:translateY(0)} }
-  @keyframes ad-spin     { to{transform:rotate(360deg)} }
-  @keyframes ad-pulse    { 0%,100%{opacity:.4} 50%{opacity:.85} }
-
-  .ad-stat { animation: ad-fade-up 0.4s ease both; }
-  .ad-stat:nth-child(1) { animation-delay: 0.05s; }
-  .ad-stat:nth-child(2) { animation-delay: 0.10s; }
-  .ad-stat:nth-child(3) { animation-delay: 0.15s; }
-  .ad-stat:nth-child(4) { animation-delay: 0.20s; }
-  .ad-stat:nth-child(5) { animation-delay: 0.25s; }
-  .ad-stat:nth-child(6) { animation-delay: 0.30s; }
-
-  .ad-section { animation: ad-fade-up 0.5s ease 0.2s both; }
-  .ad-row:hover { background: #f5f0e8 !important; }
-
-  .ad-skeleton {
-    background: #ede9df;
-    border-radius: 10px;
-    animation: ad-pulse 1.5s ease infinite;
-  }
-
-  .risk-input {
-    width: 100%;
-    padding: 9px 13px;
-    border: 1.5px solid #e8e2d6;
-    border-radius: 9px;
-    font-size: 13px;
-    color: #1a1714;
-    background: #faf7f2;
-    font-family: 'DM Sans', sans-serif;
-    outline: none;
-    transition: border-color .2s;
-  }
-  .risk-input:focus {
-    border-color: #8a7d5e;
-    box-shadow: 0 0 0 3px rgba(138,125,94,.13);
-  }
-`;
-
-// ─── Palette ──────────────────────────────────────────────────────────────────
-const C = {
-  bg:      "#f2ede4",
-  surface: "#faf7f2",
-  dark:    "#1a1714",
-  gold:    "#c9b87a",
-  goldL:   "#e8d9a0",
-  green:   "#8a7d5e",
-  muted:   "#9a8c6e",
-  border:  "#e8e2d6",
-  text:    "#1a1714",
-  textSub: "#6b6340",
-  textMut: "#b0a890",
-};
-
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-const fmtMoney = (v) =>
-  v != null ? `€${Number(v).toLocaleString("de-DE")}` : "—";
-const fmtDate = (d) =>
-  d ? new Date(d).toLocaleDateString("en-GB", { day: "2-digit", month: "2-digit", year: "numeric" }) : "—";
-const fmtRelative = (d) => {
-  if (!d) return "—";
-  const diff  = Date.now() - new Date(d).getTime();
-  const mins  = Math.floor(diff / 60000);
-  const hours = Math.floor(diff / 3600000);
-  const days  = Math.floor(diff / 86400000);
-  if (mins < 60)  return `${mins}m ago`;
-  if (hours < 24) return `${hours}h ago`;
-  return `${days}d ago`;
-};
-
-// ─── Status configs ───────────────────────────────────────────────────────────
-const LEAD_STATUS = {
-  NEW:         { color: "#d97706", bg: "#fffbeb", dot: "#d97706" },
-  IN_PROGRESS: { color: "#2563eb", bg: "#eff6ff", dot: "#2563eb" },
-  DONE:        { color: "#059669", bg: "#ecfdf5", dot: "#059669" },
-  REJECTED:    { color: "#dc2626", bg: "#fef2f2", dot: "#dc2626" },
-  DECLINED:    { color: "#64748b", bg: "#f1f5f9", dot: "#64748b" },
-};
-const CONTRACT_STATUS = {
-  PENDING_SIGNATURE: { color: "#d97706", bg: "#fffbeb" },
-  ACTIVE:            { color: "#059669", bg: "#ecfdf5" },
-  ENDED:             { color: "#64748b", bg: "#f1f5f9" },
-  CANCELLED:         { color: "#dc2626", bg: "#fef2f2" },
-};
-const APP_STATUS = {
-  PENDING:   { color: "#d97706", bg: "#fffbeb" },
-  APPROVED:  { color: "#059669", bg: "#ecfdf5" },
-  REJECTED:  { color: "#dc2626", bg: "#fef2f2" },
-  CANCELLED: { color: "#64748b", bg: "#f1f5f9" },
-};
-const PROP_EMOJI = {
-  APARTMENT: "🏢", HOUSE: "🏠", VILLA: "🏡",
-  LAND: "🌿", COMMERCIAL: "🏬", OFFICE: "🏛️",
-};
-
-const RISK_CFG = {
-  LOW:      { color: "#059669", bg: "#ecfdf5", bar: "#059669" },
-  MEDIUM:   { color: "#d97706", bg: "#fffbeb", bar: "#d97706" },
-  HIGH:     { color: "#ea580c", bg: "#fff7ed", bar: "#ea580c" },
-  CRITICAL: { color: "#dc2626", bg: "#fef2f2", bar: "#dc2626" },
-};
-
-// ─── Shared components ────────────────────────────────────────────────────────
-function StatusPill({ label, config }) {
-  const s = config || { color: "#64748b", bg: "#f1f5f9" };
-  return (
-    <span style={{
-      background: s.bg, color: s.color,
-      padding: "3px 10px", borderRadius: 20,
-      fontSize: 11, fontWeight: 600, letterSpacing: "0.3px", whiteSpace: "nowrap",
-    }}>{label}</span>
-  );
-}
-
-function StatCard({ icon, label, value, accent = C.gold, sub, onClick }) {
-  return (
-    <div className="ad-stat ad-card ad-btn" onClick={onClick}
-      style={{ padding: "18px 20px", display: "flex", alignItems: "center", gap: 14, cursor: onClick ? "pointer" : "default" }}>
-      <div style={{
-        width: 46, height: 46, borderRadius: 12, flexShrink: 0,
-        background: `${accent}18`, border: `1.5px solid ${accent}28`,
-        display: "flex", alignItems: "center", justifyContent: "center", fontSize: 20,
-      }}>{icon}</div>
-      <div style={{ flex: 1, minWidth: 0 }}>
-        <p style={{ margin: 0, fontSize: 10, fontWeight: 600, color: C.textMut, textTransform: "uppercase", letterSpacing: "0.9px", marginBottom: 3 }}>{label}</p>
-        <p style={{ margin: 0, fontSize: 26, fontWeight: 700, color: C.text, fontFamily: "'Cormorant Garamond',Georgia,serif", letterSpacing: "-0.5px", lineHeight: 1 }}>{value ?? "—"}</p>
-        {sub && <p style={{ margin: "4px 0 0", fontSize: 11, color: C.muted }}>{sub}</p>}
-      </div>
-    </div>
-  );
-}
-
-function SectionHeader({ title, count, action, onAction }) {
-  return (
-    <div style={{
-      padding: "14px 20px", borderBottom: `1px solid ${C.border}`,
-      display: "flex", alignItems: "center", justifyContent: "space-between",
-      background: "#fdf9f4",
-    }}>
-      <div style={{ display: "flex", alignItems: "center", gap: 10 }}>
-        <p style={{ margin: 0, fontSize: 15, fontWeight: 700, color: C.text, fontFamily: "'Cormorant Garamond',Georgia,serif" }}>{title}</p>
-        {count != null && (
-          <span style={{ background: `${C.gold}22`, color: C.textSub, padding: "2px 9px", borderRadius: 20, fontSize: 11, fontWeight: 600 }}>{count}</span>
-        )}
-      </div>
-      {action && (
-        <button className="ad-btn" onClick={onAction}
-          style={{ padding: "6px 14px", borderRadius: 9, background: C.dark, color: "#f5f0e8", fontSize: 11.5, fontWeight: 500 }}>
-          {action}
-        </button>
-      )}
-    </div>
-  );
-}
-
-function EmptyRow({ icon, text }) {
-  return (
-    <div style={{ padding: "36px 20px", textAlign: "center", color: C.textMut }}>
-      <div style={{ fontSize: 30, marginBottom: 8 }}>{icon}</div>
-      <p style={{ fontSize: 13, margin: 0 }}>{text}</p>
-    </div>
-  );
-}
-
-function Skeleton({ rows = 3, h = 44 }) {
-  return (
-    <div style={{ padding: "14px 16px", display: "flex", flexDirection: "column", gap: 8 }}>
-      {Array.from({ length: rows }).map((_, i) => (
-        <div key={i} className="ad-skeleton" style={{ height: h, opacity: 1 - i * 0.15 }} />
-      ))}
-    </div>
-  );
-}
-
-function Toast({ msg, type = "success", onDone }) {
-  useEffect(() => { const t = setTimeout(onDone, 3200); return () => clearTimeout(t); }, [onDone]);
-  return (
-    <div style={{
-      position: "fixed", bottom: 26, right: 26, zIndex: 9999,
-      background: C.dark, color: type === "error" ? "#f09090" : "#90c8a8",
-      padding: "11px 18px", borderRadius: 12, fontSize: 13,
-      boxShadow: "0 10px 36px rgba(0,0,0,0.32)",
-      border: `1px solid ${type === "error" ? "rgba(240,128,128,0.15)" : "rgba(144,200,168,0.15)"}`,
-      fontFamily: "'DM Sans',sans-serif", animation: "ad-toast 0.2s ease",
-      display: "flex", alignItems: "center", gap: 8,
-    }}>
-      {type === "error" ? "⚠️" : "✅"} {msg}
-    </div>
-  );
-}
-
-// ─── AI Payment Risk Panel — inline ──────────────────────────────────────────
-function PaymentRiskPanel({ clientId }) {
-  const [result,  setResult]  = useState(null);
-  const [loading, setLoading] = useState(false);
-  const [error,   setError]   = useState(null);
-
-  useEffect(() => {
-    let cancelled = false;
-    const analyze = async () => {
-      setLoading(true); setError(null);
-      try {
-        const r = await api.get(`/api/ai/payments/risk/${clientId}`);
-        if (!cancelled) setResult(r.data);
-      } catch(e) {
-        if (!cancelled) setError(e.response?.data?.message || "Error analyzing risk");
-      } finally {
-        if (!cancelled) setLoading(false);
-      }
-    };
-    analyze();
-    return () => { cancelled = true; };
-  }, [clientId]);
-
-  const cfg = result ? (RISK_CFG[result.risk_level] || RISK_CFG.MEDIUM) : null;
-
-  if (loading) return (
-    <div style={{textAlign:"center",padding:"24px 0",color:C.textMut,fontSize:13}}>
-      <div style={{width:20,height:20,border:"2px solid #e8e2d6",borderTop:`2px solid ${C.gold}`,borderRadius:"50%",animation:"ad-spin .7s linear infinite",margin:"0 auto 10px"}}/>
-      Analyzing payment risk...
-    </div>
-  );
-
-  if (error) return (
-    <div style={{padding:"12px 14px",background:"#fef2f2",border:"1.5px solid #fca5a5",borderRadius:10,fontSize:13,color:"#dc2626",marginTop:8}}>
-      ⚠️ {error}
-    </div>
-  );
-
-  if (!result) return null;
-
-  return (
-    <div style={{marginTop:4,animation:"ad-fade-up .3s ease"}}>
-      <div style={{marginBottom:14}}>
-        <div style={{display:"flex",alignItems:"center",justifyContent:"space-between",marginBottom:6}}>
-          <span style={{fontSize:13,fontWeight:600,color:C.text}}>Risk Score</span>
-          <span style={{background:cfg.bg,color:cfg.color,padding:"3px 12px",borderRadius:20,fontSize:12,fontWeight:700}}>
-            {result.risk_level} ({result.risk_score}/10)
-          </span>
-        </div>
-        <div style={{height:8,background:"#e8e2d6",borderRadius:10,overflow:"hidden"}}>
-          <div style={{height:"100%",width:`${result.risk_score*10}%`,background:cfg.bar,borderRadius:10,transition:"width .5s ease"}}/>
-        </div>
-      </div>
-      <div style={{display:"grid",gridTemplateColumns:"1fr 1fr",gap:10,marginBottom:14}}>
-        <div style={{background:"#f5f0e8",borderRadius:10,padding:"10px 14px",textAlign:"center"}}>
-          <p style={{margin:0,fontSize:10,color:C.muted,textTransform:"uppercase",letterSpacing:"0.7px",marginBottom:3}}>Total Payments</p>
-          <p style={{margin:0,fontSize:20,fontWeight:700,color:C.text,fontFamily:"'Cormorant Garamond',Georgia,serif"}}>{result.total_payments}</p>
-        </div>
-        <div style={{background:"#fef2f2",borderRadius:10,padding:"10px 14px",textAlign:"center"}}>
-          <p style={{margin:0,fontSize:10,color:"#9a3412",textTransform:"uppercase",letterSpacing:"0.7px",marginBottom:3}}>Overdue</p>
-          <p style={{margin:0,fontSize:20,fontWeight:700,color:"#dc2626",fontFamily:"'Cormorant Garamond',Georgia,serif"}}>{result.overdue_payments}</p>
-        </div>
-      </div>
-      <div style={{background:cfg.bg,borderRadius:10,padding:"10px 14px",marginBottom:10}}>
-        <p style={{margin:"0 0 4px",fontSize:10.5,color:cfg.color,fontWeight:600,textTransform:"uppercase"}}>Analysis</p>
-        <p style={{margin:0,fontSize:13,color:C.text}}>{result.reasoning}</p>
-      </div>
-      <div style={{background:"#eff6ff",borderRadius:10,padding:"10px 14px"}}>
-        <p style={{margin:"0 0 4px",fontSize:10.5,color:"#2563eb",fontWeight:600,textTransform:"uppercase"}}>Recommendation</p>
-        <p style={{margin:0,fontSize:13,color:C.text}}>{result.recommendation}</p>
-      </div>
-    </div>
-  );
-}
-
-// ─── AI Search wrapper ────────────────────────────────────────────────────────
-function AiPaymentRiskClientSearch() {
-  const [inputVal, setInputVal] = useState("");
-  const [clientId, setClientId] = useState(null);
-
-  const handleAnalyze = () => {
-    const id = parseInt(inputVal, 10);
-    if (id > 0) setClientId(id);
-  };
-
-  const handleClear = () => {
-    setClientId(null);
-    setInputVal("");
-  };
-
-  return (
-    <div>
-      <div style={{display:"flex",gap:10,marginBottom:16,maxWidth:420}}>
-        <input
-          className="risk-input"
-          type="number"
-          min="1"
-          placeholder="Enter client ID..."
-          value={inputVal}
-          onChange={e => setInputVal(e.target.value)}
-          onKeyDown={e => e.key === "Enter" && handleAnalyze()}
-        />
-        <button
-          className="ad-btn"
-          onClick={handleAnalyze}
-          disabled={!inputVal}
-          style={{
-            padding:"9px 18px", borderRadius:9,
-            background: inputVal ? C.dark : "#e8e2d6",
-            color: inputVal ? C.goldL : C.textMut,
-            fontSize:13, fontWeight:600,
-            border:`1px solid ${inputVal ? C.gold+"40" : "transparent"}`,
-            flexShrink:0,
-          }}
-        >
-          ✨ Analyze
-        </button>
-        {clientId && (
-          <button
-            className="ad-btn"
-            onClick={handleClear}
-            style={{
-              padding:"9px 14px", borderRadius:9,
-              border:`1.5px solid ${C.border}`,
-              background:"transparent", color:C.textSub,
-              fontSize:13, fontWeight:500, flexShrink:0,
-            }}
-          >
-            Clear
-          </button>
-        )}
-      </div>
-      {clientId && <PaymentRiskPanel key={clientId} clientId={clientId} />}
-    </div>
-  );
-}
+import { C, LEAD_STATUS, CONTRACT_STATUS, APP_STATUS, PROP_EMOJI, fmtMoney, fmtDate, fmtRelative } from "../../components/agent/dashboard/dashboardConstants";
+import { CSS } from "../../components/agent/dashboard/dashboardStyles";
+import { StatusPill, StatCard, SectionHeader, EmptyRow, Skeleton, Toast } from "../../components/agent/dashboard/dashboardComponents";
+import AiPaymentRiskClientSearch from "../../components/agent/dashboard/AiPaymentRiskClientSearch";
 
 // ═══════════════════════════════════════════════════════════════════════════════
 export default function AgentDashboard() {
